% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullMixtureRegressionFunctions.R
\name{BICFMR}
\alias{BICFMR}
\title{BICFMR computes the BIC of the FMR model}
\usage{
BICFMR(data, hatTheta)
}
\arguments{
\item{data}{the data (d columns for X and the last on e for Y)}

\item{hatTheta}{a list containing  the maximum likelihood estimation}
}
\value{
the BIC criterio  -2 l(\theta; Y) +2 log (n) * (K-1 + (d+1) K + d K + d (d+1) /2 *K )
}
\description{
BICFMR computes the BIC of the FMR model
}

