% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullMixtureRegressionFunctions.R
\name{drawFMRParam}
\alias{drawFMRParam}
\title{drawFMRParam draws a set of parameter for the FMR model}
\usage{
drawFMRParam(K, d, p = 3)
}
\arguments{
\item{K}{the number of cluster}

\item{d}{the dimension for X}

\item{parameters}{the parameter that is a list with component
pik the mixture proportion
foreach k in 1,\ldots, K mu_k, \Sigma_k, \beta_k and \sigma_k^Y
if parameters == NULL, then the set of parameters is drawn}
}
\value{
a list with Y, X the observations, Z and parameters the list of parameters
}
\description{
drawFMRParam draws a set of parameter for the FMR model
}

