% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationPredictionFunctions.R
\name{estimateMixtureModel}
\alias{estimateMixtureModel}
\title{estimateMixtureModel uses Rmixmod To estimate a d dimensionnal mixture model and save the results with the appropriate fileName}
\usage{
estimateMixtureModel(data, Kmax = 5, crit = "ICL",
  fileName = "ClusterData", model = "all", resDir = "")
}
\arguments{
\item{data}{the data to be fitted}

\item{Kmax}{the maximum number of components}

\item{crit}{the criterion to be maximixed for the selection of the number of components}

\item{fileName}{the name uses to store the results}

\item{model}{the kind of model to be tested}

\item{ResDir}{the optionnal directory name to store the results}
}
\value{
a Rmixmod results
}
\description{
estimateMixtureModel uses Rmixmod To estimate a d dimensionnal mixture model and save the results with the appropriate fileName
}

