% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationPredictionFunctions.R
\name{estimateXonlyMixtureModel}
\alias{estimateXonlyMixtureModel}
\title{estimateMixtureModel uses Rmixmod To estimate a d dimensionnal mixture model and save the results with the appropriate fileName}
\usage{
estimateXonlyMixtureModel(data, Kmax = 5, crit = "ICL",
  fileName = "ClusterXData", model = "all", resDir = "")
}
\arguments{
\item{data}{the data to be fitted, the slast column being the data to be predicted}

\item{Kmax}{the maximum number of components}

\item{crit}{the criterion to be maximixed for the selection of the number of components}

\item{fileName}{the name uses to store the results}

\item{model}{the kind of model to be tested}

\item{ResDir}{the optionnal directory name to store the results}
}
\value{
a list of the BestModel for clustering x and the regresssion coef
}
\description{
estimateMixtureModel uses Rmixmod To estimate a d dimensionnal mixture model and save the results with the appropriate fileName
}

