% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullMixtureRegressionFunctions.R
\name{initEMFullMixtureRegression}
\alias{initEMFullMixtureRegression}
\title{initEMFullMixtureRegression uses Rmixmod  and regression to provide good initialisations for the FMR EM algorithm.}
\usage{
initEMFullMixtureRegression(data, K)
}
\arguments{
\item{data}{the data to be fitted}

\item{K}{the number of cluster}
}
\value{
a list of parameters with
 pik the proportion of the mixture,
 sigmaX a list of K covariance matrix,
 muX  a list of K d dimensionnal expectation and
 theta a list of K regression parameters
 sigmaY a list of K residual variance for the regression
}
\description{
initEMFullMixtureRegression uses Rmixmod  and regression to provide good initialisations for the FMR EM algorithm.
}

