% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullMixtureRegressionFunctions.R
\name{logLikeFMR}
\alias{logLikeFMR}
\title{logLikeFMR computes the log likelihood of the FMR model}
\usage{
logLikeFMR(data, K, currentParameters)
}
\arguments{
\item{data}{the data (d columns for X and the last on e for Y)}

\item{K}{the number of cluster}

\item{currentParameters}{a list containing theta}
}
\value{
the loglikelihood l(\theta; Y)
}
\description{
logLikeFMR computes the log likelihood of the FMR model
}

