% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullMixtureRegressionFunctions.R
\name{predFMR}
\alias{predFMR}
\title{predictFMR computes the prediction for the given covariates data with  FMR model and parameter theta}
\usage{
predFMR(covariates, theta)
}
\arguments{
\item{covariates}{a matrix or data.frame with the d columns of X}

\item{theta}{a list containing  the parameter value to be used for prediction}
}
\value{
a vector of length nrow(covariates) containing the prediction i.e \hat{Y^{(k)}}_i \P(Z_i=k\vert X_i)
}
\description{
predictFMR computes the prediction for the given covariates data with  FMR model and parameter theta
}

