% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationPredictionFunctions.R
\name{predictionXonly}
\alias{predictionXonly}
\title{predictionXonly predicts Xd given the estimated parameters of the mixture and the regression coef within component X1:(d-1)}
\usage{
predictionXonly(XonlyModel, covariate, Y = NULL)
}
\arguments{
\item{XonlyModel}{a list containing bestModel, the mixture model to create the group and regCoef the regression coef within group}

\item{covariate}{the vector (or matrix (n,d-1) ) X1:(d-1) of the covariates to predict Xd}

\item{Y}{an optionnel vector with length equals to the number of rows for Xd. If given, the rmse is computed.}
}
\value{
a numeric vector Xd size(n)
}
\description{
predictionXonly predicts Xd given the estimated parameters of the mixture and the regression coef within component X1:(d-1)
}

