% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullMixtureRegressionFunctions.R
\name{simFMR}
\alias{simFMR}
\title{ssimFMR draws sample from model FMR in the paper. That is a mixiture where given component k
X_i\vert Zik=1 is a normal distribution of mean mu_k and variance \Sigma_k and given X_i and Z_ik Y is normal
distribution with mean t(X_i) \beta_k and variance \sigma^Y_k}
\usage{
simFMR(n, K, parameters = NULL, seed = NULL, d = 2)
}
\arguments{
\item{n}{the number of observations}

\item{K}{the number of cluster}

\item{parameters}{the parameter for the FMR model, that is a list with component
pik the mixture proportion
foreach k in 1,\ldots, K muX_k, \SigmaX_k, \beta_k and \sigmaY_k
if parameters == NULL, then the set of parameters is drawn}

\item{d}{the dimension for X}
}
\value{
a list with Y, X the observations, Z and parameters the list of parameters
}
\description{
ssimFMR draws sample from model FMR in the paper. That is a mixiture where given component k
X_i\vert Zik=1 is a normal distribution of mean mu_k and variance \Sigma_k and given X_i and Z_ik Y is normal
distribution with mean t(X_i) \beta_k and variance \sigma^Y_k
}

