% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullMixtureRegressionFunctions.R
\name{updateFMRParameters}
\alias{updateFMRParameters}
\title{updateParameters runs one iteration of the EM algorithm}
\usage{
updateFMRParameters(data, K, currentParameters)
}
\arguments{
\item{data}{the data (d columns for X and the last on e for Y)}

\item{K}{the number of cluster}

\item{currentParameters}{the current state of parameters (theta')}
}
\value{
the result of the optimisation of Q(theta, theta')
}
\description{
updateParameters runs one iteration of the EM algorithm
}

